% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequalVarShink.R
\name{a_hat_mle}
\alias{a_hat_mle}
\title{Likelihood function of the James-Stein shrinkage factor.}
\usage{
a_hat_mle(stat, vars, a_hat)
}
\arguments{
\item{stat}{Input statistics to be shrinkage estimated.}

\item{vars}{Corresponding variances of equal length.}

\item{a_hat}{Shrinkage intensity to be estimated.}
}
\value{
The likelihood of the function given the parameters.
}
\description{
To be used in MLE computation of the James-Stein shrinkage factor.
}
\references{
http://projecteuclid.org/euclid.ss/1331729986
}

