% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDups.R
\name{allDups}
\alias{allDups}
\title{Identify all duplicates values in a vector.}
\usage{
allDups(x)
}
\arguments{
\item{x}{The input vector.}
}
\value{
A logical vector.
}
\description{
By default the base R function duplicated only identifies the duplicated values after the first in a vector as TRUE. This function identifies all of the duplicates as true.
}

