% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbio_package.R
\docType{package}
\name{bayesbio}
\alias{bayesbio}
\alias{bayesbio-package}
\title{bayesbio: Miscellaneous functions useful in bioinformatics and Bayesian statistics}
\description{
A hodgepodge of hopefully helpful functions. Two of these perform
shrinkage estimation: one using a simple weighted method where the user can
specify the degree of shrinkage required, and one using James-Stein shrinkage
estimation for the case of unequal variances.
}

