% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindFill.R
\name{cbindFill}
\alias{cbindFill}
\title{cbind while converting missing entries to NA.}
\usage{
cbindFill(...)
}
\arguments{
\item{...}{A set of vectors separated by commas.}
}
\value{
A matrix that combines the inputted vectors.
}
\description{
cbind usually malfunctions on vector of unequal lengths; this function allows vectors of unequal length to be combined, while filling the missing entries with NAs.
}
\references{
http://r.789695.n4.nabble.com/How-to-join-matrices-of-different-row-length-from-a-list-td3177212.html; http://stackoverflow.com/a/7962286/560791
}

