% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStrings.R
\name{createStrings}
\alias{createStrings}
\title{Creates random, unique character strings.}
\usage{
createStrings(number, length, upper = FALSE)
}
\arguments{
\item{number}{Specifies the number of character strings that should be created.}

\item{length}{Specifies the length of each character string in letters.}

\item{upper}{Binary parameter specifying whether the character strings should be uppercase. Default = FALSE, so the character strings are all lowercase.}
}
\description{
Makes them unique by randomly choosing the character strings; and, in case it is necessary, adding numbers to the end using make.unique.
}
\references{
http://stackoverflow.com/a/1439541/560791
}

