% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggHorizBar.R
\name{ggHorizBar}
\alias{ggHorizBar}
\title{Create a color-labeled horizontal bar plot in ggplot2.}
\usage{
ggHorizBar(data_df, dataCol, namesCol, labelsCol, decreasing = TRUE)
}
\arguments{
\item{data_df}{Data frame with columns to specify the data values, the row names, and the fill colors of each of the bars.}

\item{dataCol}{The column name that specifies the values to be plotted.}

\item{namesCol}{The column name that specifies the corresponding names for each of the bar plots to be plotted.}

\item{labelsCol}{The column name that specifies the groups of the labels.}

\item{decreasing}{Logical specifying whether the values in dataCol should be in decreasing order.}
}
\value{
A ggplot2 object, which can be plotted via the plot() function or saved via the ggsave() function.
}
\description{
This function takes a data frame and creates a horizontal (by default) bar plot from it while ordering the values.
}

