% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMatSym.R
\name{makeMatSym}
\alias{makeMatSym}
\title{Replace the upper or lower triangle of a matrix with the other to make it symmetric.}
\usage{
makeMatSym(mat, replaceUpper = TRUE)
}
\arguments{
\item{mat}{The matrix to be made symmetric.}

\item{replaceUpper}{Whether the upper triangle of the matrix should be replaced by the lower triangle. Default = TRUE; if FALSE, the lower triangle of the matrix is replaced by the upper triangle.}
}
\value{
A matrix that has been made symmetric.
}
\description{
The replaced values will be lost following the operation of this function.
}

