% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multiple pattern gsub.}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{Character vector of patterns to match.}

\item{replacement}{Character vector of replacements for each pattern.}

\item{x}{Character vector in which the gsub should be performed.}

\item{...}{Additional arguments to grep.}
}
\description{
An extension to gsub that handles vectors of patterns and replacements, avoiding recursion problems associated with overlap at the extense of computation time.
}
\references{
http://stackoverflow.com/a/15254254/560791
}

