% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.adjust.nlp.R
\name{p.adjust.nlp}
\alias{p.adjust.nlp}
\title{Adjust p-values where n is less than p.}
\usage{
p.adjust.nlp(p, method = p.adjust.methods, n = length(p))
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{method}{Correction method.}

\item{n}{Number of comparisons to be made.}
}
\description{
This function recapitulates p.adjust but allows the number of hypothesis tests n to be less than the number of p-values p. Statistical properties of the p-value adjustments may not hold.
}
\references{
http://stackoverflow.com/a/30110186/560791
}

