% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_error.R
\name{std_error}
\alias{std_error}
\title{Find the standard error of the sampling distribution of a statistic.}
\usage{
std_error(x, na.rm = TRUE)
}
\arguments{
\item{x}{The numeric vector whose standard error should be calculated.}

\item{na.rm}{Logical; TRUE indicates that NAs should be removed from the vector prior to calculating the standard error, and vice versa for FALSE.}
}
\value{
A one-element numeric vector giving the standard error.
}
\description{
Finds the standard error of a numeric vector (i.e., the standard deviation divided by the square root of the sample size); by default, removes NAs prior to calculation.
}

