% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supsubDiag.R
\name{subsupDiag}
\alias{subsupDiag}
\title{Add values to the super- and sub-diagonals of a matrix.}
\usage{
subsupDiag(matrix, x)
}
\arguments{
\item{matrix}{Matrix whose super- and sub-diagonals values should be replaced.}

\item{x}{Numeric vector used to replace values in the matrix. If the inputted vector is not of the same length as both the super- and sub-diagonals of the matrix, then short vector recycling will occur (e.g., x can be one value to replace all of the super- and sub-diagonals of the matrix with that one value).}
}
\value{
The original matrix with the values added.
}
\description{
Takes a matrix and adds values to the values that are one above the diagonal (ie the superdiagonal) and the values that are one below the diagonal (ie the subdiagonal).
}
\references{
http://stackoverflow.com/a/9885186/560791
}

