% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequalVarShink.R
\name{unequalVarShrink}
\alias{unequalVarShrink}
\title{Perform James-Stein shrinkage estimation using unequal variances}
\usage{
unequalVarShrink(stat, vars, verbose = TRUE)
}
\arguments{
\item{stat}{Input statistics to be shrinkage estimated.}

\item{vars}{Corresponding variances of equal length.}

\item{verbose}{Whether information about the algorithm should be reported.}
}
\value{
A data frame containing the shrinkage estimated statistics.
}
\description{
Traditional JS shrinkage estimation assumes equal variances for each of the data points, while this algorithm extends JS shrinkage estimation to entries with different variances.
}
\references{
http://projecteuclid.org/euclid.ss/1331729986
}

